Fixed a bug in ChessBoard implementation. First complete implementation on selfplay loop.
Bug desc:
Previously didn't clear dark piece bitboard when a dark piece was captured. This was discovered when getting a frustrating amount of sigsegvs out of my selfplay initialization, then I discovered that's due to mismatch between chessboard state and dark piece state. A lot of work with gdb (mainly getting symbols to work) finally told me that the problem belongs to the most basic component.